feat: First commit.
Adds code for installable package `pyfavor'
`pyfavor' is a library for automating the process of executing WDN
simulations in WNTR and creating input Excel files to eFAVOR -
a software package written in C# for burst detection by monitoring
flows across the network under different pressure settings.

The code can be used as a collection of functions, e.g. in scripts
writen in Jupyter notebooks or as a command-line interface.